[김태민] 에딧 수정

+++ FinalProject/src/main/webapp/WEB-INF/mypage/myPageEdit.jsp
@@ -13,9 +13,41 @@ $(function(){
 	
 	var passchecklist1=true;
 	var passchecklist2=true;
+	var passchecklist3=true;
+//	var passchecklist4=true;
 	var hpchecklist=true;
 
 
+// 비밀번호 값이 입력되어있는지 확인
+
+/*$(".userpasswd3").blur(function(){
+
+	var p5 = $("#pass3").val();
+	    
+	if( p5 == "" ) {
+	    passchecklist3=false;
+	    	
+	} else{
+	    passchecklist3=true;
+	}
+});*/
+	
+//비밀번호 입력값이 기존의 비밀번호와 일치하는지 확인
+
+	$(".userpasswd3").blur(function(){
+
+		var p3 = $("#pass3").val();
+	    var p4 = $("#existingpw").val();
+	    
+	    if( p3 != p4 ) {
+	    	passchecklist3=false;
+	    	
+	    } else{
+	    	passchecklist3=true;
+	    }
+	});
+	
+	
 //비밀번호 8~16자리의 문자와 숫자를 조합했는지 확인
 $(".userpasswd").blur(function(){
 
@@ -25,7 +57,7 @@ $(".userpasswd").blur(function(){
 	if(!check1){
 		
 		passchecklist1=false;
-		alert("비밀번호는 8자리 이상 16자리 이하로 문자와 숫자를 조합하여 주십시오.");
+		
 	} else{		
 		passchecklist1=true;
 	}
@@ -40,7 +72,7 @@ $(".userpasswd2").blur(function(){
     
     if( p1 != p2 ) {
     	passchecklist2=false;
-    	//alert("비밀번호가 일치하지 않습니다.");
+    	
     } else{
     	passchecklist2=true;
     }
@@ -54,7 +86,7 @@ $(".userhp").blur(function(){
 	
 	if(!hp_pattern.test(hp)) {
 		hpchecklist=false;
-		alert("전화번호는 숫자만을 입력해야 합니다.");
+		
     } else {
     	hpchecklist=true;
     }
@@ -64,31 +96,55 @@ $(".userhp").blur(function(){
 
 $(".btn-update-information").click(function(){
 
+	if(passchecklist3 == false){
+		alert("비밀번호를 확인해 주세요.");
+		return false;
+	}
 	if(passchecklist1 == false) {
-
+		alert("비밀번호는 8자리 이상 16자리 이하로 문자와 숫자를 조합하여 주십시오.");
 		return false;
 	}
 	if(passchecklist2 == false){
-
+		alert("변경 될 비밀번호가 일치하지 않습니다.");
 		return false;
 	}
 	if(hpchecklist == false){
-		
+		alert("전화번호는 숫자만을 입력해야 합니다.");
 		return false;
 	}
+	
+
+	
 });
 
 
 
 });
+
 function memDelete() {
 		var a=confirm("정말 탈퇴하시겠습니까?")
-		
+		var p3 = $("#pass3").val();
+	    var p4 = $("#existingpw").val();
+	    
+	    
 		if(a){
 			
-			location.href="/mypage/deleteedit"
+			if(p3 == p4){
+			
+			return location.href="/mypage/deleteedit"
+			}
+			else{
+				
+				alert("비밀번호를 확인해 주십시오.");
+				return false
+			}
+				
 		}
-}
+		else{
+			
+			return false
+		}
+};
 
 </script>
 </head>
@@ -115,6 +171,7 @@ function memDelete() {
 			<form action="/mypage/updateedit" method="post" class="form-memberform">
 				<input type="hidden" name="allchecklist" value = "allchecklist" >
 	  			<input type="hidden" name="mnum" value="${dto.mnum }">
+				<input type="hidden" name="existingpw" id="existingpw" value="${dto.password }">
 				
 					<table class="edit-content">
 						<tr>
@@ -143,17 +200,23 @@ function memDelete() {
 							<td><input type="text" name="phone" class="form-control-inputbox userhp"
 								placeholder="010-0000-1324" value="${dto.phone}"></td>
 						</tr>
-
+						
 						<tr>
 							<th align="left">비밀번호</th>
-							<td><input type="password" name="password" id="pass"
-								class="form-control-inputbox userpasswd" placeholder="비밀번호를 입력해주세요." required="required"></td>
+							<td><input type="password" name="password" id="pass3"
+								class="form-control-inputbox userpasswd3" placeholder="비밀번호를 입력해주세요." value= "" required = required ></td>
+						</tr>
+
+						<tr>
+							<th align="left">변경할 비밀번호</th>
+							<td><input type="password" name="password1" id="pass"
+								class="form-control-inputbox userpasswd" placeholder="변경할 비밀번호를 입력해주세요." value= ""></td>
 						</tr>
 						
 						<tr>
 							<th align="left">비밀번호 확인</th>
 							<td><input type="password" name="passcheck" id="pass2"
-								class="form-control-inputbox userpasswd2" placeholder="비밀번호를 다시 입력해 주세요." required="required"></td>
+								class="form-control-inputbox userpasswd2" placeholder="변경할 비밀번호를 다시 입력해 주세요." ></td>
 						</tr>
 						
 					</table>
